+++ example/templatepack/base/metadata.rb
@@ -1,7 +1,7 @@
-name             '|{ cookbook["name"] }|'
+name             '|{ cookbook['name'] }|'
 maintainer       'Rackspace US, Inc.'
 maintainer_email 'rackspace-cookbooks@rackspace.com'
 license          'All rights reserved'
-description      'Installs/Configures |{ cookbook["name"] }|'
+description      'Installs/Configures |{ cookbook['name'] }|'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'