+++ tests/templatepack/base/recipes/default.rb
@@ -1,6 +1,0 @@
-#
-# Cookbook Name:: |{ cookbook['name'] }|
-# Recipe :: default
-#
-# Copyright |{ cookbook['year'] }|, Rackspace
-#