+++ tests/templatepack/base/test/unit/spec/default_spec.rb
@@ -1,7 +1,0 @@
-require_relative 'spec_helper'
-
-describe '|{ cookbook['name'] }|::default' do
-  let(:chef_run) do
-    ChefSpec::Runner.new.converge('|{ cookbook['name'] }|::default')
-  end
-end